default methods to avoid interface changes
